Added support for unions
